web_tools/pyproject.toml
[project]
name = "agent-tool-new"
version = "0.1.0"
description = "Agent tool for web scraping and LLM interactions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "uvicorn>=0.37.0",
    "openai>=1.109.1",
    "httpx>=0.28.1",
    "aiohttp>=3.12.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.4.0",
    "python-dotenv==1.0.1",
    "pydantic>=2.11.9",
    "asyncio>=3.4.3",
    "numpy>=2.3.3",
    "pandas>=2.2.3",
    "ddgs==9.4.0",
    "redis>=6.4.0",
    "prometheus-client>=0.23.1",
    "pillow>=11.3.0",
    "crawl4ai==0.7.4",
]
